exercises/function_app/get_time/get_time.py
import logging
import datetime
import azure.functions as func

def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')

    return func.HttpResponse(f'{str(datetime.datetime.now())}')