exercises/function_app/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-keyvault
azure-storage-blob
azure-identity